pyproject.toml
[project]
name = "datapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.67.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "telethon>=1.42.0",
]
